docs(swagger): add swagger build script
