+++ requirements.txt
@@ -1,4 +1,2 @@
 Flask==1.0.2
 requests==2.18.4
-flasgger==0.9.1
-PyYAML==3.13
